Updated the syllabus
Just see if the link for principles of recursion is correct. It has got the reference of trees which might be not known to people. Although this reference can be neglected for the time being, but still just had a doubt. I also need to brush up my concept of recursion, which I will do asap dw. For this particular concept I think we should do a meet-up with everyone.